+++ python/fate_llm/algo/inferdpt/utils.py
@@ -1,0 +1,2 @@
+import fate_llm
+from fate_llm.algo.inferdpt import inferdpt
